+++ src/generator/writeReport.ts
@@ -3,6 +3,7 @@ import { CompilerContext } from "../context";
 import { PackageFileFormat } from "../packaging/fileFormat";
 import { getType } from "../types/resolveDescriptors";
 import { Writer } from "../utils/Writer";
+import { TypeDescription } from "../types/types";
 
 export function writeReport(ctx: CompilerContext, pkg: PackageFileFormat) {
     const w = new Writer();
@@ -43,6 +44,54 @@ export function writeReport(ctx: CompilerContext, pkg: PackageFileFormat) {
     Object.entries(abi.errors!).forEach(([t, abiError]) => {
         w.write(`${t}: ${abiError.message}`);
     });
+    w.append();
+
+    const t = getType(ctx, pkg.name);
+    const writtenEdges: Set<string> = new Set();
+
+    // Trait Inheritance Diagram
+    w.write(`# Trait Inheritance Diagram`);
+    w.append();
+    w.write("```mermaid");
+    w.write("graph TD");
+    function writeTraits(t: TypeDescription) {
+        for (const trait of t.traits) {
+            const edge = `${t.name} --> ${trait.name}`;
+            if (writtenEdges.has(edge)) {
+                continue;
+            }
+            writtenEdges.add(edge);
+            w.write(edge);
+            writeTraits(trait);
+        }
+    }
+    w.write(t.name);
+    writeTraits(t);
+    w.write("```");
+    w.append();
+
+    writtenEdges.clear();
+
+    // Contract Dependency Diagram
+    w.write(`# Contract Dependency Diagram`);
+    w.append();
+    w.write("```mermaid");
+    w.write("graph TD");
+    function writeDependencies(t: TypeDescription) {
+        for (const dep of t.dependsOn) {
+            const edge = `${t.name} --> ${dep.name}`;
+            if (writtenEdges.has(edge)) {
+                continue;
+            }
+            writtenEdges.add(edge);
+            w.write(edge);
+            writeDependencies(dep);
+        }
+    }
+    writtenEdges.clear();
+    w.write(t.name);
+    writeDependencies(t);
+    w.write("```");
 
     return w.end();
 }