+++ src/grammar/ast.ts
@@ -679,7 +679,7 @@ export function __DANGER_resetNodeId() {
     nextId = 1;
 }
 
-// Test equality of ASTExpressions.
+// Test equality of AstExpressions.
 export function eqExpressions(
     ast1: AstExpression,
     ast2: AstExpression,
@@ -689,78 +689,81 @@ export function eqExpressions(
     }
 
     try {
-    switch (ast1.kind) {
-        case "null":
-            return true;
-        case "boolean":
-            return ast1.value === (ast2 as AstBoolean).value;
-        case "number":
-            return ast1.value === (ast2 as AstNumber).value;
-        case "string":
-            return ast1.value === (ast2 as AstString).value;
-        case "id":
-            return eqNames(ast1, ast2 as AstId);
-        case "method_call":
-            return (
-                eqNames(ast1.method, (ast2 as AstMethodCall).method) &&
-                eqExpressions(ast1.self, (ast2 as AstMethodCall).self) &&
-                eqExpressionArrays(ast1.args, (ast2 as AstMethodCall).args)
-            );
-        case "init_of":
-            return (
-                eqNames(ast1.contract, (ast2 as AstInitOf).contract) &&
-                eqExpressionArrays(ast1.args, (ast2 as AstInitOf).args)
-            );
-        case "op_unary":
-            return (
-                ast1.op === (ast2 as AstOpUnary).op &&
-                eqExpressions(ast1.operand, (ast2 as AstOpUnary).operand)
-            );
-        case "op_binary":
-            return (
-                ast1.op === (ast2 as AstOpBinary).op &&
-                eqExpressions(ast1.left, (ast2 as AstOpBinary).left) &&
-                eqExpressions(ast1.right, (ast2 as AstOpBinary).right)
-            );
-        case "conditional":
-            return (
-                eqExpressions(
-                    ast1.condition,
-                    (ast2 as AstConditional).condition,
-                ) &&
-                eqExpressions(
-                    ast1.thenBranch,
-                    (ast2 as AstConditional).thenBranch,
-                ) &&
-                eqExpressions(
-                    ast1.elseBranch,
-                    (ast2 as AstConditional).elseBranch,
-                )
-            );
-        case "struct_instance":
-            return (
-                eqNames(ast1.type, (ast2 as AstStructInstance).type) &&
-                eqParameterArrays(ast1.args, (ast2 as AstStructInstance).args)
-            );
-        case "field_access":
-            return (
-                eqNames(ast1.field, (ast2 as AstFieldAccess).field) &&
-                eqExpressions(
-                    ast1.aggregate,
-                    (ast2 as AstFieldAccess).aggregate,
-                )
-            );
-        case "static_call":
-            return (
-                eqNames(ast1.function, (ast2 as AstStaticCall).function) &&
-                eqExpressionArrays(ast1.args, (ast2 as AstStaticCall).args)
-            );
+        switch (ast1.kind) {
+            case "null":
+                return true;
+            case "boolean":
+                return ast1.value === (ast2 as AstBoolean).value;
+            case "number":
+                return ast1.value === (ast2 as AstNumber).value;
+            case "string":
+                return ast1.value === (ast2 as AstString).value;
+            case "id":
+                return eqNames(ast1, ast2 as AstId);
+            case "method_call":
+                return (
+                    eqNames(ast1.method, (ast2 as AstMethodCall).method) &&
+                    eqExpressions(ast1.self, (ast2 as AstMethodCall).self) &&
+                    eqExpressionArrays(ast1.args, (ast2 as AstMethodCall).args)
+                );
+            case "init_of":
+                return (
+                    eqNames(ast1.contract, (ast2 as AstInitOf).contract) &&
+                    eqExpressionArrays(ast1.args, (ast2 as AstInitOf).args)
+                );
+            case "op_unary":
+                return (
+                    ast1.op === (ast2 as AstOpUnary).op &&
+                    eqExpressions(ast1.operand, (ast2 as AstOpUnary).operand)
+                );
+            case "op_binary":
+                return (
+                    ast1.op === (ast2 as AstOpBinary).op &&
+                    eqExpressions(ast1.left, (ast2 as AstOpBinary).left) &&
+                    eqExpressions(ast1.right, (ast2 as AstOpBinary).right)
+                );
+            case "conditional":
+                return (
+                    eqExpressions(
+                        ast1.condition,
+                        (ast2 as AstConditional).condition,
+                    ) &&
+                    eqExpressions(
+                        ast1.thenBranch,
+                        (ast2 as AstConditional).thenBranch,
+                    ) &&
+                    eqExpressions(
+                        ast1.elseBranch,
+                        (ast2 as AstConditional).elseBranch,
+                    )
+                );
+            case "struct_instance":
+                return (
+                    eqNames(ast1.type, (ast2 as AstStructInstance).type) &&
+                    eqParameterArrays(
+                        ast1.args,
+                        (ast2 as AstStructInstance).args,
+                    )
+                );
+            case "field_access":
+                return (
+                    eqNames(ast1.field, (ast2 as AstFieldAccess).field) &&
+                    eqExpressions(
+                        ast1.aggregate,
+                        (ast2 as AstFieldAccess).aggregate,
+                    )
+                );
+            case "static_call":
+                return (
+                    eqNames(ast1.function, (ast2 as AstStaticCall).function) &&
+                    eqExpressionArrays(ast1.args, (ast2 as AstStaticCall).args)
+                );
+        }
+    } catch (e) {
+        // In principle, the assertions "as Ast___" should not fail
+        // because ast1 and ast2 have the same kind inside the switch.
+        return false;
     }
-} catch (e) {
-    // In principle, the assertions "as Ast___" should not fail 
-    // because ast1 and ast2 have the same kind inside the switch.
-    return false;
-}
 }
 
 function eqParameters(